Продолжение шинковки

+++ Runaway/Objects/ShipStats.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Runtime.CompilerServices;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -15,10 +16,10 @@ namespace Runaway.Objects
             BulletSpeed = 10;
             FirePower = 1;
         }
-        public static int HP { get; set; }
-        public static int LvlHP { get => HP - 20 / 10 + 1; }        
-        public static int WeaponDamage { get; set; }
-        public static int LvlWeaponDamage { get => WeaponDamage - 5 / 5 + 1; }
+        public static long HP { get; set; }
+        public static long LvlHP { get => HP - 20 / 10 + 1; }        
+        public static long WeaponDamage { get; set; }
+        public static long LvlWeaponDamage { get => WeaponDamage - 5 / 5 + 1; }
         public static int BulletSpeed { get; set; }
         public static int LvlBulletSpeed { get => Math.Abs(BulletSpeed - 10) / 1 + 1; }
         public static int MaxBulletSpeed { get => 1; }
